requirements.txt
requests==2.31.0
flask[async]==2.3.3
pandas
python-dotenv==1.0.0
datasets
transformers
librosa
evaluate
jiwer
gradio
langchain
langchain-experimental
openai
pymysql
deep_translator
faster_whisper
cryptography
pinecone-client